Move SardanaBuffer unit tests to its own module
